운동 계획 게시글 작성/편집 패널 중복 제거 및 메서드화 시도
